+++ src/test/kotlin/framework/test-util.kt
@@ -40,6 +40,7 @@ import com.intellij.testFramework.LexerTestCase
 import com.intellij.testFramework.fixtures.JavaCodeInsightTestFixture
 import com.intellij.util.ReflectionUtil
 import org.junit.jupiter.api.Assertions
+import org.opentest4j.AssertionFailedError
 
 typealias ProjectBuilderFunc =
     ProjectBuilder.(path: String, code: String, configure: Boolean, allowAst: Boolean) -> VirtualFile
@@ -130,4 +131,47 @@ fun testInspectionFix(fixture: JavaCodeInsightTestFixture, basePath: String, fix
     val intention = fixture.findSingleIntention(fixName)
     fixture.launchAction(intention)
     fixture.checkResult(expected)
+}
+
+fun <T> assertEqualsUnordered(expected: Collection<T>, actual: Collection<T>) {
+    val expectedSet = expected.toSet()
+    val actualSet = actual.toSet()
+    val notFound = expectedSet.minus(actualSet)
+    val notExpected = actualSet.minus(expectedSet)
+
+    if (notExpected.isNotEmpty() && notFound.isNotEmpty()) {
+        val message = """|
+      |Expecting actual:
+      |  $actual
+      |to contain exactly in any order:
+      |  $expected
+      |elements not found:
+      |  $notFound
+      |and elements not expected:
+      |  $notExpected
+    """.trimMargin()
+        throw AssertionFailedError(message, expected, actual)
+    }
+    if (notFound.isNotEmpty()) {
+        val message = """|
+      |Expecting actual:
+      |  $actual
+      |to contain exactly in any order:
+      |  $expected
+      |but could not find the following elements:
+      |  $notFound
+    """.trimMargin()
+        throw AssertionFailedError(message, expected, actual)
+    }
+    if (notExpected.isNotEmpty()) {
+        val message = """|
+      |Expecting actual:
+      |  $actual
+      |to contain exactly in any order:
+      |  $expected
+      |but the following elements were unexpected:
+      |  $notExpected
+    """.trimMargin()
+        throw AssertionFailedError(message, expected, actual)
+    }
 }